requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0

# LLM API integrations
openai==1.3.0
anthropic==0.21.0  # If using Claude models

# Data processing
pandas==2.1.0
numpy==1.24.3

# Email handling
pydantic==2.4.2
email-validator==2.0.0

# Web scraping (if needed)
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1

# Utilities
tqdm==4.66.1  # Progress bars
python-dateutil==2.8.2